move totalRatings, averageRating, recommendedCount, and percentageRecommended calculations into ./SharedComponents/sharedFns.js
